Guard against None or empty string data folders.
This commit limits the possible values for the data_folder variables
in the AwsService, DataRequest, and DownloadRequest classes to
non-empty strings. This fixes the error thrown when providing no value
for the data_folder argument in AwsProductRequest, and it enforces a
consistent state across the classes where data_folder always holds a
value that can be used with os.path.join.

Note that joining the empty string with another path is an idempotent
operation, whereas joining with the current directory ('.') results
in a new path that explicitly references "this directory". That is,
joining the empty string with 'bob' results in 'bob', but joining
the current directory with 'bob' results in './bob'. The two result
paths are equivalent as long as any further path joining happens
via os.path.join (which it should).