Use XDG_CONFIG_HOME for config folder
Retrieve the default folder for storing config files from the XDG_CONFIG_HOME env var if present. Fall back to "~/.config" if the env var is not declared.

This allows customization of the config folder. For example when running on a YARN cluster, ~ expands to /home/ and attempting to save the config raises a permission error.